+++ tests/models/charge_collection/test_simple_persistence.py
@@ -10,13 +10,7 @@ import typing as t
 import numpy as np
 import pytest
 
-from pyxel.detectors import (
-    CMOS,
-    CMOSCharacteristics,
-    CMOSGeometry,
-    Environment,
-    Material,
-)
+from pyxel.detectors import CMOS, CMOSCharacteristics, CMOSGeometry, Environment
 from pyxel.models.charge_collection import simple_persistence
 
 
@@ -31,7 +25,6 @@ def cmos_5x10() -> CMOS:
             pixel_vert_size=10.0,
             pixel_horz_size=10.0,
         ),
-        material=Material(),
         environment=Environment(),
         characteristics=CMOSCharacteristics(),
     )