+++ tests/models/photon_generation/test_load_image.py
@@ -26,7 +26,6 @@ from pyxel.detectors import (
     CCDCharacteristics,
     CCDGeometry,
     Environment,
-    Material,
     ReadoutProperties,
 )
 from pyxel.models.photon_generation import load_image
@@ -59,7 +58,6 @@ def ccd_10x10() -> CCD:
             pixel_vert_size=10.0,
             pixel_horz_size=10.0,
         ),
-        material=Material(),
         environment=Environment(),
         characteristics=CCDCharacteristics(),
     )